reinitialisation de score lors de regenerate , alert de confirmation de delete ,redirection vers questions apres generations  et quelques autres ameliorations

+++ frontend/src/components/Myquizzes.js
@@ -44,33 +44,53 @@ const Myquizzes = () => {
   }, [user]);
 
   const deleteFormResponse = async (Id) => {
-    try {
-      await axios.delete(`http://localhost:5000/api/responseForm/${Id}`, {
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': `Bearer ${user.token}`
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'Do you really want to delete this quizze? This process cannot be undone.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it !',
+      cancelButtonText: 'No, cancel !',
+      reverseButtons: true
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await axios.delete(`http://localhost:5000/api/responseForm/${Id}`, {
+            headers: {
+              'Content-Type': 'application/json',
+              'Authorization': `Bearer ${user.token}`
+            }
+          });
+          setFormResponses(prevResponses => prevResponses.filter(response => response._id !== Id));
+          Swal.fire({
+            icon: 'success',
+            title: 'Deleted',
+            text: 'Your Questions has been deleted successfully.',
+            timer: 1500,
+            showConfirmButton: false
+          });
+        } catch (error) {
+          console.error('Error deleting quiz:', error);
+          Swal.fire({
+            icon: 'error',
+            title: 'Error',
+            text: 'Failed to delete Questions. Please try again.',
+            timer: 2000,
+            showConfirmButton: false
+          });
         }
-      });
-      setFormResponses(prevResponses => prevResponses.filter(response => response._id !== Id));
-      Swal.fire({
-        icon: 'success',
-        title: 'Deleted',
-        text: 'Your Questions has been deleted successfully.',
-        timer: 1500,
-        showConfirmButton: false
-      });
-    } catch (error) {
-      console.error('Error deleting quiz:', error);
-      Swal.fire({
-        icon: 'error',
-        title: 'Error',
-        text: 'Failed to delete Questions. Please try again.',
-        timer: 2000,
-        showConfirmButton: false
-      });
-    }
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire({
+          icon: 'error',
+          title: 'Cancelled',
+          text: 'Your questions are safe.',
+          timer: 1500,
+          showConfirmButton: false
+        });
+      }
+    });
   };
-
+  
   const filteredResponses = formResponses.filter(response => 
     response.topic.toLowerCase().includes(searchQuery.toLowerCase())
   );
